fix: harden bms-model against overflow, negative zero, and sentinel values
Change Note::play_time() return type from i32 to i64 to prevent
wrapping for songs over ~35.8 minutes, consistent with other time
accessor methods.

Canonicalize negative zero in f64_to_key to prevent BTreeMap misordering
when -0.0 enters through float arithmetic. Deduplicate private f64_key
to delegate to the canonical implementation.

Add sentinel guard (-1 check) in osu_decoder before using mapping values
as lane indices, preventing silent note drops on unmapped columns.

Add .max(0) guard on nlane_len in total_notes_full to prevent negative
i32 wrapping to huge usize, and bounds check on scratch_key loop.

+++ crates/bms-model/src/note.rs
@@ -154,8 +154,8 @@ impl Note {
         self.data_mut().duration = duration;
     }
 
-    pub fn play_time(&self) -> i32 {
-        (self.data().playtime / 1000) as i32
+    pub fn play_time(&self) -> i64 {
+        self.data().playtime / 1000
     }
 
     pub fn milli_play_time(&self) -> i64 {
